refactor(api): split constants
