+++ src/api/httpCodes.js
@@ -1,0 +1,4 @@
+/**
+ * API OK status code
+ */
+export const HTTP_OK = 200;
