+++ src/api/mutations/constants.js
@@ -7,8 +7,3 @@ export const MUTATION_LOGIN = 'login';
  * SignUp muatation name
  */
 export const MUTATION_SIGNUP = 'signup';
-
-/**
- * API OK status code
- */
-export const HTTP_OK = 200;
